+++ dbtenv/__init__.py
@@ -14,9 +14,10 @@ import sys
 from typing import Any, List, Optional
 
 
-__version__ = '1.1.1'
+__version__ = '1.2.0'
 
-VENVS_DIRECTORY = os.path.normpath('~/.dbt/versions')
+DEFAULT_VENVS_DIRECTORY = os.path.normpath('~/.dbt/versions')
+DEFAULT_VENVS_PREFIX    = ''
 
 GLOBAL_VERSION_FILE = os.path.normpath('~/.dbt/version')
 LOCAL_VERSION_FILE  = '.dbt_version'
@@ -26,7 +27,10 @@ AUTO_INSTALL_VAR          = 'DBTENV_AUTO_INSTALL'
 DEBUG_VAR                 = 'DBTENV_DEBUG'
 DEFAULT_INSTALLER_VAR     = 'DBTENV_DEFAULT_INSTALLER'
 PYTHON_VAR                = 'DBTENV_PYTHON'
+QUIET_VAR                 = 'DBTENV_QUIET'
 SIMULATE_RELEASE_DATE_VAR = 'DBTENV_SIMULATE_RELEASE_DATE'
+VENVS_DIRECTORY_VAR       = 'DBTENV_VENVS_DIRECTORY'
+VENVS_PREFIX_VAR          = 'DBTENV_VENVS_PREFIX'
 
 
 def string_is_true(value: str) -> bool:
@@ -71,9 +75,13 @@ class Installer(Enum):
 
 
 class Version(distutils.version.LooseVersion):
-    def __init__(self, version: str, source: Optional[str] = None) -> None:
+    def __init__(self, version: str, source: Optional[str] = None, source_description: Optional[str] = None) -> None:
         self.pypi_version = self.homebrew_version = self.raw_version = version.strip()
         self.source = source
+        if source and not source_description:
+            self.source_description = f"set by {source}"
+        else:
+            self.source_description = source_description
 
         version_match = re.match(r'(?P<version>\d+\.\d+\.\d+)(-?(?P<prerelease>[a-z].*))?', self.raw_version)
         self.is_semantic = version_match is not None
@@ -138,7 +146,9 @@ class Environment:
         self.project_file = self.find_file_along_working_path('dbt_project.yml')
         self.project_directory = os.path.dirname(self.project_file) if self.project_file else None
 
-        self.venvs_directory     = os.path.expanduser(VENVS_DIRECTORY)
+        self.venvs_directory = os.path.expanduser(self.env_vars.get(VENVS_DIRECTORY_VAR) or DEFAULT_VENVS_DIRECTORY)
+        self.venvs_prefix = self.env_vars.get(VENVS_PREFIX_VAR) or DEFAULT_VENVS_PREFIX
+
         self.global_version_file = os.path.expanduser(GLOBAL_VERSION_FILE)
 
         self.homebrew_installed = False
@@ -166,7 +176,32 @@ class Environment:
     @debug.setter
     def debug(self, value: bool) -> None:
         self._debug = value
-        LOGGER.setLevel(logging.DEBUG if self._debug else logging.INFO)
+        self.update_logging_level()
+
+    _quiet: Optional[bool] = None
+
+    @property
+    def quiet(self) -> bool:
+        if self._quiet is None:
+            if QUIET_VAR in self.env_vars:
+                self._quiet = string_is_true(self.env_vars[QUIET_VAR])
+            else:
+                self._quiet = False
+
+        return self._quiet
+
+    @quiet.setter
+    def quiet(self, value: bool) -> None:
+        self._quiet = value
+        self.update_logging_level()
+
+    def update_logging_level(self) -> None:
+        if self.debug:
+            LOGGER.setLevel(logging.DEBUG)
+        elif self.quiet:
+            LOGGER.setLevel(logging.ERROR)
+        else:
+            LOGGER.setLevel(logging.INFO)
 
     _default_installer: Optional[Installer] = None